% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{conditions}
\alias{conditions}
\alias{alert}
\alias{warn}
\alias{stop}
\title{Conditions}
\usage{
alert(...)

warn(..., .demo = FALSE)

stop(..., .demo = FALSE)
}
\description{
These functions provide a way for signaling conditions. \code{rui::warn()} and
\code{rui::stop()} are drop-in replacements for \code{\link[base:warning]{base::warning()}} and
\code{\link[base:stop]{base::stop()}}, which support \strong{glue} strings. The function \code{rui::alert()}
can be used for alerts with \strong{cli} inline styles as well. We recommend
doing so before issuing a warning or error.
}
