% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{console}
\alias{console}
\title{Title}
\usage{
console(..., object = NULL, .envir = parent.frame())
}
\arguments{
\item{...}{Character vectors supporting \strong{glue} strings and \strong{cli} inline
styles.}

\item{object}{Object to print the {rui} way through rui::inspect().}

\item{.envir}{}
}
\description{
Title
}
