% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R, R/deprecated.R
\name{multi-line-feedback}
\alias{multi-line-feedback}
\alias{entitle}
\alias{inform}
\alias{approve}
\alias{disapprove}
\alias{title}
\title{Provide multi-line feedback}
\usage{
entitle(...)

inform(...)

approve(...)

disapprove(...)

title(...)
}
\arguments{
\item{...}{Character vectors supporting \strong{glue} strings and \strong{cli} inline
styles.}
}
\description{
These functions provide a way to provide multi-line feedback. This is
typically useful for any kind of small tasks you want to provide feedback
about, or longer tasks for which you need the feedback to persist in the
console (\emph{e.g.} running external code that does provide essential
information (in such a case, these are useful in combination with
\code{\link[processx:run]{processx::run()}} and its \code{stdout_line_callback} argument)) Use
\itemize{
\item \code{rui::entitle()} for naming sections,
\item \code{rui::inform()} for providing information,
\item \code{rui::approve()} for succesful completion of a task or a positive test,
and
\item \code{rui::disapprove()} for unsuccesful completion or a negative test.
}
}
\seealso{
\code{\link[glue:glue]{glue::glue()}}, \code{\link[cli:inline-markup]{cli::inline-markup}}
}
