% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R, R/deprecated.R
\name{object-inspection}
\alias{object-inspection}
\alias{expose}
\alias{display}
\alias{inspect}
\alias{extract}
\alias{show}
\title{Object inspection}
\usage{
expose(..., level = 1)

display(...)

inspect(object, levels = 1)

extract(...)

show(...)
}
\arguments{
\item{...}{Character vectors supporting \strong{glue} strings and \strong{cli} inline
styles.}

\item{level}{Level of indentation, typically useful for (nested) lists.
Defaults to 1.}

\item{object}{Object to print the structure of.}

\item{levels}{Number of levels to include. Defaults to 1.}
}
\description{
These functions provide a means to inspect the internal structure of
objects, or design a UI to do so, which is consistent with the rest of the
{rui} functionality. \code{rui::expose()} can be used to print information on
different parts of an object (where the \code{.} in the prefix should be
familiar to users of the {magrittr} pipe, and the \code{$} is one of the
subsetting operators), whereas \code{rui::display()} would be more appropriate for
the atomic types. \code{rui::inspect()} attempts to use both these functions
together with \code{rui::entitle()} to process the output of a \code{utils::str()} call,
and print the structure consistent with the rest of the {rui} functions,
borrowing some style elements from \code{tibble:::print.tbl()} and
\code{tibble::glimpse()}.
}
\seealso{
\code{\link[glue:glue]{glue::glue()}}, \code{\link[cli:inline-markup]{cli::inline-markup}}
}
