% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{standard-text}
\alias{standard-text}
\alias{tell}
\title{Output standard text messages}
\usage{
tell(...)
}
\arguments{
\item{...}{Character vectors supporting \strong{glue} strings and \strong{cli} inline
styles.}
}
\description{
Output standard text messages
}
\seealso{
\code{\link[glue:glue]{glue::glue()}}, \code{\link[cli:inline-markup]{cli::inline-markup}}
}
