% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{user-interaction}
\alias{user-interaction}
\alias{copy}
\alias{do}
\alias{ask}
\title{User interaction}
\usage{
copy(x, copy = rlang::is_interactive(), .envir = parent.frame())

do(...)

ask(...)
}
\arguments{
\item{...}{Character vectors supporting \strong{glue} strings and \strong{cli} inline
styles.}
}
\description{
These functions provide a way to interact with the user. Use
\itemize{
\item \code{rui::copy()} to print and copy R code to the clipboard,
\item \code{rui::do()} to list a to do item,
\item \code{rui::ask()} to ask a yes/no question.
Note \code{rui::copy()} does not support \strong{cli} styles at the moment.
}
}
\seealso{
\code{\link[glue:glue]{glue::glue()}}, \code{\link[cli:inline-markup]{cli::inline-markup}}
}
