% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R, R/deprecated.R
\name{conditions}
\alias{conditions}
\alias{alert}
\alias{warn}
\alias{error}
\alias{stop}
\title{Conditions}
\usage{
alert(..., warn = FALSE, error = FALSE)

warn(..., .demo = FALSE)

error(..., .demo = FALSE)

stop(...)
}
\arguments{
\item{...}{Character vectors supporting \strong{glue} strings and \strong{cli} inline
styles.}

\item{warn}{Logical. Should a warning be issued with the same message.}

\item{error}{Logical. Should an error be issued with the same message.}
}
\description{
These functions provide a way for signaling conditions. \code{rui::warn()} and
\code{rui::error()} are drop-in replacements for \code{\link[base:warning]{base::warning()}} and
\code{\link[base:stop]{base::stop()}}, which support \strong{glue} strings and \strong{cli} inline styles.
ANSI colours are lost however. For retaining colours, \code{rui::alert()}
can be used. We recommend doing so before issuing a warning or error.
}
\seealso{
\code{\link[glue:glue]{glue::glue()}}, \code{\link[cli:inline-markup]{cli::inline-markup}}
}
